rename test case, cleaning up debug statements, adding tests on state names with other special characters
