--- tests/asl_CallCenter/settings.json
@@ -1,4 +1,0 @@
-{
-    "workflow_name": "__cc_test",
-    "workflow_description_file": "workflow_cc_test.json"
-}
